fix discussions
