+++ app/helpers/operations_helper.rb
@@ -11,11 +11,7 @@ module OperationsHelper
     "#{amount} #{CURRENCIES.fetch(currency.downcase.to_sym, currency.upcase)}"
   end
 
-  def button_text
-    case controller.action_name
-    when 'new' then 'Save'
-    when 'edit' then 'Edit'
-    else 'Submit'
-    end
+  def button
+    button_text
   end
 end
