+++ app/helpers/users_helper.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
 module UsersHelper
-  def button_text
-    case controller.action_name
-    when 'new' then 'Save'
-    when 'edit' then 'Edit'
-    else 'Submit'
-    end
+  def button
+    button_text
   end
 end
