+++ db/migrate/20220804080913_create_users.rb
@@ -4,8 +4,9 @@ class CreateUsers < ActiveRecord::Migration[7.0]
   def change
     create_table :users do |t|
       t.string :name, null: false
-      t.string :email, null: false
+      t.string :email, null: false, unique: true
       t.string :password, null: false
+      t.boolean :is_deleted, default: false
 
       t.timestamps
     end
