--- tableIncomAndExpeditureWithERB_development
@@ -1,2 +1,0 @@
-ls
-;
